First take on help generator.
The help generator takes a git repository of markdown files and
turns it into a set of static HTML pages. A template of choice
can be used to render the pages.

The goal is to make it as easy as possible to maintain help content,
while creating a nice integrated online help for use as part of
a web site or web application.